Refactor: 뉴스 도메인 관련 테스트 코드의 참조 위치 수정 #252
